+++ Tests/Unit/Authentication/BackEndLoginManagerTest.php
@@ -6,6 +6,7 @@ namespace OliverKlee\Oelib\Tests\Unit\Authentication;
 
 use Nimut\TestingFramework\TestCase\UnitTestCase;
 use OliverKlee\Oelib\Authentication\BackEndLoginManager;
+use OliverKlee\Oelib\Model\BackEndUser;
 
 /**
  * @covers \OliverKlee\Oelib\Authentication\BackEndLoginManager
@@ -58,5 +59,39 @@ class BackEndLoginManagerTest extends UnitTestCase
 
         // @phpstan-ignore-next-line We explicitly check for contract violations here.
         $this->subject->getLoggedInUser('');
+    }
+
+    /**
+     * @test
+     */
+    public function getLoggedInUserUidWithoutLoginReturnsZero(): void
+    {
+        self::assertSame(0, $this->subject->getLoggedInUserUid());
+    }
+
+    /**
+     * @test
+     */
+    public function getLoggedInUserUidWithLoginReturnsUserUid(): void
+    {
+        $uid = 12;
+        $user = new BackEndUser();
+        $user->setUid($uid);
+
+        $this->subject->setLoggedInUser($user);
+
+        self::assertSame($uid, $this->subject->getLoggedInUserUid());
+    }
+
+    /**
+     * @test
+     */
+    public function getLoggedInUserUidWithLoginWithoutUIdReturnsZeo(): void
+    {
+        $user = new BackEndUser();
+
+        $this->subject->setLoggedInUser($user);
+
+        self::assertSame(0, $this->subject->getLoggedInUserUid());
     }
 }